WIP: consolidate routes and handlers
the whole PathBuilder thing might be avoidable if we instead put the GetParam
and CreatePathForRoute functions into the api package.
