+++ atc/api/buildserver/create.go
@@ -32,7 +32,7 @@ func (s *Server) CreateBuild(team db.Team) http.Handler {
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusCreated)
 
-		err = json.NewEncoder(w).Encode(present.Build(build))
+		err = json.NewEncoder(w).Encode(present.Build(build, s.router))
 		if err != nil {
 			hLog.Error("failed-to-encode-build", err)
 			w.WriteHeader(http.StatusInternalServerError)
