+++ atc/api/buildserver/get.go
@@ -15,7 +15,7 @@ func (s *Server) GetBuild(build db.Build) http.Handler {
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusOK)
 
-		err := json.NewEncoder(w).Encode(present.Build(build))
+		err := json.NewEncoder(w).Encode(present.Build(build, s.router))
 		if err != nil {
 			logger.Error("failed-to-encode-build", err)
 			w.WriteHeader(http.StatusInternalServerError)
