+++ atc/api/jobserver/create_build.go
@@ -80,7 +80,7 @@ func (s *Server) CreateJobBuild(pipeline db.Pipeline) http.Handler {
 			}
 		}
 
-		err = json.NewEncoder(w).Encode(present.Build(build))
+		err = json.NewEncoder(w).Encode(present.Build(build, s.router))
 		if err != nil {
 			logger.Error("failed-to-encode-build", err)
 			w.WriteHeader(http.StatusInternalServerError)
