+++ atc/api/pipelineserver/list_builds.go
@@ -82,7 +82,7 @@ func (s *Server) ListPipelineBuilds(pipeline db.Pipeline) http.Handler {
 		atc := make([]atc.Build, len(builds))
 		for i := 0; i < len(builds); i++ {
 			build := builds[i]
-			atc[i] = present.Build(build)
+			atc[i] = present.Build(build, s.router)
 		}
 
 		err = json.NewEncoder(w).Encode(atc)
