+++ atc/api/resourceserver/check.go
@@ -70,7 +70,7 @@ func (s *Server) CheckResource(dbPipeline db.Pipeline) http.Handler {
 
 		w.WriteHeader(http.StatusCreated)
 
-		err = json.NewEncoder(w).Encode(present.Build(build))
+		err = json.NewEncoder(w).Encode(present.Build(build, s.router))
 		if err != nil {
 			logger.Error("failed-to-encode-check", err)
 			w.WriteHeader(http.StatusInternalServerError)
