+++ tsa/deleter_test.go
@@ -3,6 +3,7 @@ package tsa_test
 import (
 	"context"
 
+	"github.com/concourse/concourse/atc/routes"
 	"github.com/concourse/concourse/tsa"
 
 	"code.cloudfoundry.org/lager/lagerctx"
@@ -32,7 +33,7 @@ var _ = Describe("Deleter", func() {
 
 		fakeATC = ghttp.NewServer()
 
-		atcEndpoint := atc.NewEndpoint(fakeATC.URL())
+		atcEndpoint := routes.NewEndpoint(fakeATC.URL())
 
 		token := &oauth2.Token{TokenType: "Bearer", AccessToken: "yo"}
 		httpClient := oauth2.NewClient(oauth2.NoContext, oauth2.StaticTokenSource(token))
