+++ tsa/worker_status_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 
+	"github.com/concourse/concourse/atc/routes"
 	"github.com/concourse/concourse/tsa"
 	"golang.org/x/oauth2"
 
@@ -35,7 +36,7 @@ var _ = Describe("Worker Status Test", func() {
 
 		fakeATC = ghttp.NewServer()
 
-		atcEndpoint := atc.NewEndpoint(fakeATC.URL())
+		atcEndpoint := routes.NewEndpoint(fakeATC.URL())
 
 		token := &oauth2.Token{TokenType: "Bearer", AccessToken: "yo"}
 		httpClient := oauth2.NewClient(oauth2.NoContext, oauth2.StaticTokenSource(token))
